refactor: 디렉토리 오타 수정
